requirements.txt
pytest~=7.3.1
PyAutoGUI~=0.9.54
pytesseract~=0.3.10
Pillow~=9.5.0
lxml~=4.9.2
selenium~=4.9.1
setuptools~=65.5.1
pandas~=2.0.3